requirements.txt
flask
flask_sqlalchemy
gunicorn
pyjwt